Consolidate `temp` into `tmp`
To be unix friendly ❤️
