+++ plugins/tickets/src/modules/data/blocks/ticket/reducers/__tests__/ui.test.js
@@ -3,9 +3,9 @@
  */
 import ui, { DEFAULT_STATE } from '../ui';
 import { actions } from '@moderntribe/tickets/data/blocks/ticket';
-import temp from '../temp';
+import tmp from '../tmp';
 
-jest.mock( '../temp', () => ( {
+jest.mock( '../tmp', () => ( {
 	__esModule: true,
 	default: jest.fn(),
 } ) );
@@ -46,7 +46,7 @@ describe( 'UI reducer', () => {
 
 	test( 'Tmp actions are passed down to the reducer', () => {
 		ui( DEFAULT_STATE, actions.setTempSharedCapacity( 100 ) );
-		expect( temp ).toHaveBeenCalledTimes( 1 );
-		expect( temp ).toHaveBeenCalledWith( DEFAULT_STATE.tmp, actions.setTempSharedCapacity( 100 ) );
+		expect( tmp ).toHaveBeenCalledTimes( 1 );
+		expect( tmp ).toHaveBeenCalledWith( DEFAULT_STATE.tmp, actions.setTempSharedCapacity( 100 ) );
 	} );
 } );
